requirements.txt
python-dotenv==0.21.0
pandas>=1.3.5
beautifulsoup4==4.11.0
textblob>=0.17.1